Dodanie potwierdzenia chęci usunięcia klienta.

+++ MotoFitAcademy/OpenDayApplication/Viewmodel/ClientsViewModel.cs
@@ -93,6 +93,9 @@ namespace OpenDayApplication.Viewmodel
       IsClientEditVisible = false;
       if (EditedClient != null && EditedClient.ID != 0)
       {
+        if (MessageBox.Show("Are you sure to delete this user?", "Confirm", MessageBoxButton.YesNo, MessageBoxImage.Stop) == MessageBoxResult.No)
+          return;
+
         _clientsManager.DeleteClient(EditedClient);
         RefreshClients();
       }
